% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreedyFunction.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic function that solves Knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A Data Frame with two variables(x and w)}

\item{W}{The Total weight}
}
\value{
The Value and the elements
}
\description{
Greedy heuristic function that solves Knapsack problem
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
