% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicFunction.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Computes the knapsack problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A Data Frame with two variables(x and w)}

\item{W}{The Total weight}
}
\value{
The Value and the elements
}
\description{
An approach to solve the knapsack problem using dynamic programming
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
